aws: Move unit tests from clusterinfo_test.go
